Added Speed/Damage Reduction and No Shooting Powerdowns

+++ CSE125Group2Game/Server/Player.h
@@ -48,7 +48,7 @@ class Player : public Moveable {
 
   PickupType mPickup;
   unsigned long mPickupEndTime = 0;
-  int mDamageMultiplier = 1;
+  float mDamageMultiplier = 1;
   int mEnemiesKilled;
   float mSpeedMultiplier = 1;
